Patch to fix memory bug in replying to CC messages If we're on Athena and have static krb (or other) libraries, use   them Added "athstatic" program to the release, which handles the above Cast to an int for isspace, to make gcc -Wall quiet Added 'zlist' and 'l' to basic help.

+++ editwin.c
@@ -803,7 +803,7 @@ int owl_editwin_check_dotsend(owl_editwin *e) {
       e->buff[i] = '\0';
       return(1);
     }
-    if (!isspace(e->buff[i])) {
+    if (!isspace((int) e->buff[i])) {
       return(0);
     }
   }
